third verion of collection's some Clases refactoring was performed; Collection Obejct type included.
